// src/pages/things/xpb506/wireless/Wireless2Primary/type.d.ts
export type GetWireless2PrimaryStatus =
  | ({
      result: GetWireless2PrimaryStatusResult
    } & StatusMessageType)
  | EmptyProps

export type GetWireless2PrimaryStatusResult = {
  disabled: string
  up: boolean
  name: string
  ifname: string
  device: {
    device: string
    name: string
    up: string
  }
  ssid: string
  bssid: string
  mode: string
  encryption: string
  wpa_key_mgmt: string
  is_multiple_network: boolean
  is_guest_network: boolean
  frequency: string
  channel: number
  bitrate: number
  signal: number
  quality: number
  noise: number
  link: string
  country: string
  id: string
  txpower: number
  txpoweroff: number
  wmm: {
    CWmin: WmmProps
    txoplimit: WmmProps
    CWmax: WmmProps
    aifs: WmmProps
  }
  freqlist: FreqProps[]
  assoclist: Record<
    string,
    {
      inactive: number
      rx_he: boolean
      rx_ht: boolean
      rx_mhz: number
      rx_packets: number
      rx_rate: number
      rx_vht: boolean
      tx_he: boolean
      tx_ht: boolean
      tx_mhz: number
      tx_packets: number
      tx_rate: number
      tx_vht: boolean
      tx_signal: number
      tx_noise: number
    }
  >
  name: string
  up: boolean
}

type FreqProps = {
  restrcited: string
  mhz: number
  channel: number
}

type WmmProps = {
  AC_VI: string
  AC_VO: string
  AC_BE: string
  AC_BK: string
}